resolve invenio-accounts conflict with latest version

+++ modules/invenio-accounts/invenio_accounts/sessions.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # This file is part of Invenio.
-# Copyright (C) 2015-2018 CERN.
+# Copyright (C) 2015-2024 CERN.
 #
 # Invenio is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
@@ -14,11 +14,10 @@ This module provides management functionality and populates the
 created.
 """
 
-from __future__ import absolute_import, print_function
-
 from flask import after_this_request, current_app, request, session
 from geolite2 import geolite2
 from invenio_db import db
+from invenio_rest.csrf import reset_token
 from ua_parser import user_agent_parser
 from werkzeug.local import LocalProxy
 
@@ -32,17 +31,17 @@ def _ip2country(ip):
     """Get user country."""
     if ip:
         match = geolite2.reader().get(ip)
-        return match.get('country', {}).get('iso_code') if match else None
+        return match.get("country", {}).get("iso_code") if match else None
 
 
 def _extract_info_from_useragent(user_agent):
     """Extract extra informations from user."""
     parsed_string = user_agent_parser.Parse(user_agent)
     return {
-        'os': parsed_string.get('os', {}).get('family'),
-        'browser': parsed_string.get('user_agent', {}).get('family'),
-        'browser_version': parsed_string.get('user_agent', {}).get('major'),
-        'device': parsed_string.get('device', {}).get('family'),
+        "os": parsed_string.get("os", {}).get("family"),
+        "browser": parsed_string.get("user_agent", {}).get("family"),
+        "browser_version": parsed_string.get("user_agent", {}).get("major"),
+        "device": parsed_string.get("device", {}).get("family"),
     }
 
 
@@ -56,12 +55,12 @@ def get_remote_addr():
     if not request:
         return None
 
-    address = request.headers.get('X-Real-IP', None)
+    address = request.headers.get("X-Real-IP", None)
 
     if address is None:
-        address = request.headers.get('X-Forwarded-For', request.remote_addr)
+        address = request.headers.get("X-Forwarded-For", request.remote_addr)
         if address is not None:
-            address = address.encode('utf-8').split(b',')[0].strip().decode()
+            address = address.encode("utf-8").split(b",")[0].strip().decode()
 
     return address
 
@@ -71,19 +70,21 @@ def add_session(session=None):
 
     :param session: Flask Session object to add. If None, ``session``
         is used. The object is expected to have a dictionary entry named
-        ``"user_id"`` and a field ``sid_s``
+        ``"_user_id"`` and a field ``sid_s``
     """
-    user_id, sid_s = session['user_id'], session.sid_s
+    if "_user_id" not in session and "user_id" in session:
+        # this is to support both flask-login 0.4.x as well as 0.5+,
+        # where 'user_id' was renamed to '_user_id'
+        session["_user_id"] = session["user_id"]
 
+    user_id, sid_s = session["_user_id"], session.sid_s
     with db.session.begin_nested():
         session_activity = SessionActivity(
             user_id=user_id,
             sid_s=sid_s,
             ip=get_remote_addr(),
             country=_ip2country(get_remote_addr()),
-            **_extract_info_from_useragent(
-                request.headers.get('User-Agent', '')
-            )
+            **_extract_info_from_useragent(request.headers.get("User-Agent", ""))
         )
         db.session.merge(session_activity)
 
@@ -94,6 +95,7 @@ def login_listener(app, user):
     :param app: The Flask application.
     :param user: The :class:`invenio_accounts.models.User` instance.
     """
+
     @after_this_request
     def add_user_session(response):
         """Regenerate current session and add to the SessionActivity table.
@@ -105,8 +107,10 @@ def login_listener(app, user):
         session.regenerate()
         # Save the session first so that the sid_s gets generated.
         app.session_interface.save_session(app, session, response)
-        add_session(session)
-        current_accounts.datastore.commit()
+        # Don't add impersonation sessions
+        if "_impersonator_id" not in session:
+            add_session(session)
+            current_accounts.datastore.commit()
         return response
 
 
@@ -116,14 +120,24 @@ def logout_listener(app, user):
     :param app: The Flask application.
     :param user: The :class:`invenio_accounts.models.User` instance.
     """
+
     @after_this_request
     def _commit(response=None):
-        if hasattr(session, 'sid_s'):
+        if hasattr(session, "sid_s"):
             delete_session(session.sid_s)
         # Regenerate the session to avoid session fixation vulnerabilities.
         session.regenerate()
         current_accounts.datastore.commit()
         return response
+
+
+def csrf_token_reset(app, user):
+    """Connect to the user_logged_in signal to reset the csrf token.
+
+    :param app: The Flask application.
+    :param user: The :class:`invenio_accounts.models.User` instance.
+    """
+    reset_token()
 
 
 def delete_session(sid_s):
@@ -138,8 +152,9 @@ def delete_session(sid_s):
     # Remove entries from sessionstore
     _sessionstore.delete(sid_s)
     # Find and remove the corresponding SessionActivity entry
-    with db.session.begin_nested():
-        SessionActivity.query.filter_by(sid_s=sid_s).delete()
+    if request and "_impersonator_id" not in session:
+        with db.session.begin_nested():
+            SessionActivity.query.filter_by(sid_s=sid_s).delete()
     return 1
 
 
@@ -157,6 +172,23 @@ def delete_user_sessions(user):
 
     return True
 
+
+def default_session_store_factory(app):
+    """Session store factory.
+
+    If ``ACCOUNTS_SESSION_REDIS_URL`` is set, it returns a
+    :class:`simplekv.memory.redisstore.RedisStore` otherwise
+    a :class:`simplekv.memory.DictStore`.
+    """
+    accounts_session_redis_url = app.config.get("ACCOUNTS_SESSION_REDIS_URL")
+    if accounts_session_redis_url:
+        import redis
+        from simplekv.memory.redisstore import RedisStore
+
+        return RedisStore(redis.StrictRedis.from_url(accounts_session_redis_url))
+    from simplekv.memory import DictStore
+
+    return DictStore()
 
 def session_update(app):
